backend/marketplace-svc/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.3
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.9
httpx==0.27.0
python-multipart==0.0.9
pytest==7.4.4
pytest-asyncio==0.23.5
elasticsearch-dsl==8.11.0
redis==5.0.2
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic[email]==2.6.3
asyncpg==0.29.0
aiofiles==24.1.0
aio_pika==9.5.0
pillow==11.0.0
